Remove India Server - Add Section on Moving Servers
Removed the India server due to very low traffic.  We were expecting significantly higher traffic, our assumption is that with the recent tightening of security, we have removed most of the scammers using NW to scam, and this has brought down India usage significantly as they are mostly just accessing through the scambaiter's computer.  We are actively looking at usage reports and are planning to re-implement a server in India as the effects of COVID-19 wear off and we begin to see scammers operating at full capacity again.

Also added a little common question about changing server regions.